# contracts/managed-agent/v1alpha1/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Cloud Agents Managed Agent Foundation API",
    "version": "v1alpha1"
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/admin/tenants/{tenantId}/projects": {
      "get": {
        "operationId": "adminListProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectListOrganizationId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/ProjectPageSize"
          },
          {
            "$ref": "#/components/parameters/ProjectPageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project-page.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "adminCreateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../platform/v1alpha1/schemas/project-create-request.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/admin/tenants/{tenantId}/projects/{projectId}": {
      "get": {
        "operationId": "adminGetProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/v1/tenants/{tenantId}/my-projects": {
      "get": {
        "operationId": "managedAgentListMyProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/ProjectPageSize"
          },
          {
            "$ref": "#/components/parameters/ProjectPageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project-page.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/tenants/{tenantId}/projects": {
      "get": {
        "operationId": "managedAgentListProjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectListOrganizationId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/ProjectPageSize"
          },
          {
            "$ref": "#/components/parameters/ProjectPageToken"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project-page.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "operationId": "managedAgentCreateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../../platform/v1alpha1/schemas/project-create-request.schema.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "422": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/tenants/{tenantId}/projects/{projectId}": {
      "get": {
        "operationId": "managedAgentGetProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantId"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "X-Resource-Version": {
                "$ref": "#/components/headers/ResourceVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../../platform/v1alpha1/schemas/project.schema.json"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "TenantId": {
        "name": "tenantId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
...
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
      "ProjectListOrganizationId": {
        "name": "organizationId",
        "in": "query",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
      "ProjectPageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/page-size.schema.json"
        }
      },
      "ProjectPageToken": {
        "name": "pageToken",
        "in": "query",
        "required": false,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/page-token.schema.json"
        }
      },
...
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/identifier.schema.json"
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/idempotency-key.schema.json"
        }
      },
...
    },
    "headers": {
      "ResourceVersion": {
        "required": true,
        "schema": {
          "$ref": "../../common/v1alpha1/schemas/resource-version.schema.json"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Problem",
        "headers": {
          "WWW-Authenticate": {
            "description": "Bearer authentication challenge returned with 401 responses.",
            "schema": {
              "$ref": "./schemas/www-authenticate.schema.json"
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "../../common/v1alpha1/schemas/problem.schema.json"
            }
          }
        }
      }
    }
  },
...
}